+++ docs/history.rst
@@ -26,6 +26,9 @@ Not yet released.
 Bug Fixes
 ^^^^^^^^^
 
+* Remove resources attached to packages that don't exist. (This was a
+  regression in 0.5.)
+* Warn on failure to annotate a package. (This was a regression in 0.5.)
 * Building embedded Python resources now emits warnings when ``__file__``
   is seen. (This was a regression in 0.5.)
 * Missing parent packages are now automatically added when constructing